还是很多，都是在完善细节

+++ MadgwickAHRS.h
@@ -20,7 +20,7 @@
 
 extern volatile float beta;				// algorithm gain
 extern volatile float q0, q1, q2, q3;	// quaternion of sensor frame relative to auxiliary frame
-extern volatile uint8_t sampleInterval;
+extern volatile float sampleInterval;
 extern Quaternion quat;
 //---------------------------------------------------------------------------------------------------
 // Function declarations
